+++ openmc/deplete/__init__.py
@@ -10,7 +10,7 @@ from .chain import *
 from .operator import *
 from .reaction_rates import *
 from .atom_number import *
-from .results import *
+from .stepresult import *
 from .results_list import *
 from .integrators import *
 from . import abc
