+++ openmc/deplete/abc.py
@@ -22,7 +22,7 @@ from uncertainties import ufloat
 from openmc.lib import MaterialFilter, Tally
 from openmc.checkvalue import check_type, check_greater_than
 from openmc.mpi import comm
-from .results import StepResult
+from .stepresult import StepResult
 from .chain import Chain
 from .results_list import Results
 from .pool import deplete
