+++ openmc/deplete/stepresult.py
